+++ client/src/Components/Enter.jsx
@@ -1,17 +1,10 @@
+import { NavLink, Routes, Route } from "react-router-dom"
+import Game from "./Game"
 import { useEffect, useState } from "react"
-import styled from "styled-components"
 
-export default function Enter({ animals }) {
+export default function Enter() {
   let [count, setCount] = useState(0)
   const [messages, setMessages] = useState([])
-
-  const [chosenAnimal, setChosenAnimal] = useState({})
-  useEffect(() => {
-    if (animals.length > 0) {
-      const randomAnimal = getRandomAnimal(animals)
-      setChosenAnimal(randomAnimal)
-    }
-  }, [animals])
 
   useEffect(() => subscribe(), [messages])
 
@@ -46,43 +39,28 @@ export default function Enter({ animals }) {
     }
   }
 
-  function getRandomAnimal(array) {
-    const randomIndex = Math.floor(Math.random() * array.length)
-    const randomAnimal = array[randomIndex]
-    return randomAnimal
-  }
-
   return (
     <>
       <form onSubmit={submitMessage}>
         <input type="text" name="message" placeholder="Enter name please" />
         <br />
         <button type="reset">Cancel</button>
-        {/* if hasNameMinLength(input.message.value) {setCount...} and reset input form: value= {tagInput}; reset(input.value)? */}
-        <button
-          onClick={() => {
-            setCount(count => count + 1)
-          }}
-        >
-          Join Game
-        </button>
+        {/* if hasNameMinLength(input.message.value) {setCount...} and
+        reset input form: value= {tagInput}; reset(input.value)? or
+        if name does not have mon leghth, dann join button ausgrauen?*/}
+        <NavLink to="/game" className={({ isActive }) => (isActive ? "active" : "inactive")}>
+          <button
+            onClick={() => {
+              setCount(count => count + 1)
+            }}
+          >
+            Join
+          </button>
+        </NavLink>
+        <Routes>
+          <Route path="game" element={<Game count={count} />} />
+        </Routes>
       </form>
-      <section>{messages.length > 0 ? <p>Hi, {messages[messages.length - 1]}!</p> : ""}</section>
-      <p>Number of players: {count}</p>
-      <p>Your Animal is the following:</p>
-      {chosenAnimal && (
-        <CardStyle animal={chosenAnimal} key={chosenAnimal._id}>
-          <h4>{chosenAnimal.name}</h4>
-          <p>{chosenAnimal.type}</p>
-        </CardStyle>
-      )}
     </>
   )
 }
-
-const CardStyle = styled.div`
-  margin: 0.5rem 3rem;
-  background-color: var(--beige-day);
-  border: 2px solid var(--oliv-day);
-  padding: 0.5em;
-`
